pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "coinbase-hft"
version = "0.1.0"
description = "Production-grade high-frequency trading platform for Coinbase Advanced Trade API"
requires-python = ">=3.10"
dependencies = [
    # Async runtime
    "uvloop>=0.19.0",
    # Coinbase SDK
    "coinbase-advanced-py>=1.3.0",
    # WebSocket
    "websockets>=12.0",
    # HTTP client
    "httpx>=0.27.0",
    # Data handling
    "aiosqlite>=0.20.0",
    "redis>=5.0.0",
    # Configuration
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.0",
    # CLI
    "typer>=0.12.0",
    "rich>=13.7.0",
    # Monitoring
    "prometheus-client>=0.20.0",
    # Utilities
    "tenacity>=8.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[project.scripts]
coinbase-hft = "coinbase_hft.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["coinbase_hft"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true
